docs: add initial doxygen support
This adds a 'doc' target to make and Doxygen config and layout files
that are mostly set to defaults. The notable exception is that it
has INCLUDE_ALL=true; the alternative would be to add a /*! \file */
line to every kernel file. The output will need a bit of work to be
useful for VOLK since doxygen isn't set up to treat functions as
first-class citizens of an API. After docs are generated navigate to
/your/build/dir/html/index.html . kernel docs can be found under
files>file list>kernels>volk>any_kernel_file.h (far from optimal)