Fixes Coverity Defect: 1293620 - Buffer Overrun in `volk_profile.cc`
The issue here is that `buffer` was 128 characters long, and `found` was getting
set to `128` in some cases, which was then used to index `buffer` to set a NULL
character. Since `found` is actually an index, it has been set to `127` in the
maximum scenario, and where `found` is used as a size it is incremented by one.