Fixes Cov Defect 1311671: Dereference before NULL check.
This was an unnecessary, but completely harmless, NULL check of a pointer. The
NULL check doesn't serve any purpose since the pointer is repeatedly
dereferenced in all code paths prior to the check, and the beginning of the
function guarantees that data exists at that address. Removing the check to
reduce complexity and make Coverity happy.