volk: comment out machine selected debug printf()
This commit comments out the "Using Volk machine: <machine>" debug
message on standard out, which interferes with programs trying to use
standard out for legitimate purposes.

This modification is consistent with other commented out debug printf()
calls in the volk codebase, and should be superseded by a proper
compile-time configurable logging system for the library.