Removes compiler flags from MSVC builds
Unlike GCC, MSVC does not require /arch flags to generate intrinsics for an instruction set.  However, it the flag is present, it will attempt to optimize *all* the code for that .c file.  Unfortunately, it will even optimize code in the dynamic initializer of the class (at least MSVC 2015 will).

This will cause volk to crash when running on machine without AVX as the initializer appears to be called outside of CPU guard checks.

Tested removal of the flags and found only a minor difference in performance on an AVX2 CPU and volk_profile built on an AVX2 machine ran successfully on a SSE4.1 CPU (which it crashed on previously).

volk_profile time to execute in default settings:
MSVC Original: 62.4s
MSVC NoFlags: 66.3s
GCC   Original: 59.7s
old CPU NoFlags: 143.0sec