Fix error installing Python program with Python 3 and '\u' in path.
VOLK_PYTHON_INSTALL in CMake inserts a native path to the Python
executable using Python strings and a write call when writing an
executable for a Python program. With Python 3, strings are
interpreted as Unicode by default, so when the Python exe path
contains '\u' the command will fail with a unicode decoding error. Use a
raw string instead to prevent false interpreting of strings that look
like escapes.