Boost: build without Boost and C++17
This commit enables us to build VOLK without Boost as a dependency. If
you set the `CMAKE_CXX_STANDARD` flag to 17, the build system will not
search for Boost anymore. In this case the STL <filesystem> will be
used. This include does only work with the latest compilers e.g. GCC9
and later. It does not work with GCC7 or Clang6.
This commit might help to avoid build problems that arise with Boost
1.70.