cmake: bump minimum version to 3.8
This commit bumps the minimum required CMake version to 3.8. This is in
line with GNU Radio 3.8 requirements. Ubuntu 18.04 provides CMake 3.10.
Other distros are probably ahead of Ubuntu LTS.
Alternatively, we could bump the minimum version to 3.5.1 because this
is the version Ubuntu 16.04 ships. Though, I'd argue that `master`
should require newer versions. This would also be in line with hopefully
moving to C++17 in the near future.