Py3k: remove obsolete code
Basically, we want 2 cmake variables to be set
1. `PYTHONINTERP_FOUND` to indicate we can use Python
2. `PYTHON_EXECUTABLE` to use this specific Python version.

Previously, the code worked like this:
If `PYTHON_EXECUTABLE` is already set, e.g. by user, we just use it and `set(PYTHONINTERP_FOUND TRUE)`. This is still the same.

If we did not set `PYTHON_EXECUTABLE` we try to find it. This is where I my PR takes effect.
With `find_package(PythonInterp 3)` we search for a Py3k. This will set both cmake variables. No further action required.
Only if we can't find Py3k, we go on and search for Py2k.

Finally, if that fails as well, we try:
`find_program(PYTHON_EXECUTABLE NAMES python3 python python2 python2.7)`
This should search for executable names in order of appearance. Thus, we search for `python3` first.

Since we require Python 2.7+, we can remove
```
if(PYTHON_EXECUTABLE)
    execute_process(
        COMMAND ${PYTHON_EXECUTABLE} -B -c ""
        OUTPUT_QUIET ERROR_QUIET
        RESULT_VARIABLE PYTHON_HAS_DASH_B_RESULT
    )
    if(PYTHON_HAS_DASH_B_RESULT EQUAL 0)
        set(PYTHON_DASH_B "-B")
    endif()
endif(PYTHON_EXECUTABLE)
```
in `cmake/Modules/VolkPython.cmake

I updated `.travis.yml` as well I hope the CI system does work again with my patch.