32fc_index_max: Improve speed of AVX2 version
Improve speed by using an approach similar to the one in 32f_index_max.
See comments in volk_avx2_intrinsics.h for explanation of the algorithm.
Two variants – differing in a single line – are available. Users are
advised to run volk_profile to select the faster one for their machine.