versioning: Remove .Maint from libvolk version
We use SemVer for our releases with Major.Minor.Maint logic.
Since we want Maint releases to be 100% ABI compatible, check with ABI
compliance checker, we can drop the Maint version from
`libvolk.so.major.minor`. The idea is to make it easier for maintainers
to release a new maint version.