Remove _mm256_zeroupper() calls
`_mm256_zeroupper()` issues the `vzeroupper` instruction. Usage of this
instruction prevents performance loss when VEX and non-VEX encoded
instructions are used (see [1][2][3] for details). Manual addition of
`vzeroupper` is not required because compilers are aware of the issue
and automatically use the instruction if required. Moreover, the manual
usage leads to miscompilation (wrong code generation) with GCC 10.2 and
optimization level `-O3` (confirmed in [4]; GCC 9.3 is also possibly
affected [5]).

[1] https://stackoverflow.com/questions/41303780/why-is-this-sse-code-6-times-slower-without-vzeroupper-on-skylake
[2] https://stackoverflow.com/questions/46080327/what-is-the-penalty-of-mixing-evex-and-vex-encoded-scheme
[3] https://stackoverflow.com/questions/41819514/why-do-sse-instructions-preserve-the-upper-128-bit-of-the-ymm-registers/
[4] https://github.com/gnuradio/volk/issues/384
[5] https://github.com/gnuradio/volk/issues/382