ci: Add non-x86 Actions
This commit adds some ARM tests to our GH Actions. We should be able
to run different architectures with it. We expect the workflow to be
the same as the non-container version.

Some slight differences:
- gcc-10 does not install g++-10 as well.
- Ubuntu 20.04 is missing python3-distutils.

For the moment, we test aarch64 on Ubuntu 20.04 with Clang
and GCC. armv7 tests would be nice but cmake fails to detect the the
compiler and everything fails.