kernel: Fix typo in implementation name
The AVX2 kernels in `volk_16ic_x2_dot_prod_16ic` were named incorrectly.
Instead of `_avx2` the function names ended with `_axv2`. This commit
fixes the typo.

However, this does change two function names in files that we install.
This might require us to add a wrapper with the old name that just calls
the new, and correct, name.