relicense: Add Federico
We received the following statement from Federico
----
Hi,

I'm sending you an email as it's easier for me right now. Let me know if the text below is enough.

I, Federico 'Larroca' La Rocca, hereby resubmit all my contributions to the VOLK
project and repository under the terms of the LGPL-3.0-or-later.
My GitHub handle is git-artes.
My email addresses used for contributions are: flarroca@fing.edu.uy.

I hereby agree that contributions made by me in the past, to previous
versions of VOLK, may be re-used for inclusion in VOLK 3. I understand
that VOLK 3 will be relicensed under LGPL-3.0-or-later.

best

Federico
----

Thus, we add him to the list.

Signed-off-by: Johannes Demel <demel@uni-bremen.de>