Update AUTHORS_RESUBMITTING_UNDER_LGPL_LICENSE.md
I, Ron Economos, hereby resubmit all my contributions to the VOLK
project and respository under the terms of the LGPL-3.0-or-later. My
Github handle is drmpeg. My e-mail address used for contributions is:
w6rz@comcast.net

I hereby agree that contributions made by me in the past, to previous
versions of VOLK, may be re-used for inclusion in VOLK 3. I understand
that VOLK 3 will be relicensed under LGPL-3.0-or-later.

Signed-off-by: Ron Economos <w6rz@comcast.net>