asan: Fix volk_malloc alignment bug
ASAN (the Address Sanitizer used by GCC) requires memory allocations to
be a multiple of the alignment. To replicate the bug, use a version of
libvolk without this fix, call volk_malloc() with a number of byte
that's not a multiple of the alignment and compile it with the Address
sanitizer enable (-fsanitize=address). The software will error out
and complain about the alignement. This patch fixes it by adding the
missing number of bytes to the size variable so that it becomes a
multiple of the alignment.

Signed-off-by: AlexandreRouma <alexandre.rouma@gmail.com>