lgpl: add The Aerospace Corporation to the list
I, Kyle Logue, as a representative of The Aerospace Corporation, ("Aerospace")
on behalf of all Aerospace contributors, hereby resubmit all of our
contributions to the VOLK project and repository under the terms of the
LGPLv3+. GitHub handles associated with Aerospace contributors include but are
not limited to 'jessica-iwamoto' and 'dmiralles2009' and include all @aero.org
email addresses.

I, on behalf of all Aerospace contributors, hereby agree that contributions
made by Aerospace in the past, to previous versions of VOLK, may be re-used for
inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPLv3+.

Signed-off-by: Kyle Logue <oss@aero.org>