lgpl: Add Valerii to the list
We received the following statement

I, Valerii Zapodovnikov, hereby resubmit all my contributions to the
VOLK project and repository under the terms of the LGPL-3.0-or-later.
My GitHub handle is `ValZapod`. I hereby agree that contributions made
by me in the past, to previous versions of VOLK, may be re-used for
inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under
LGPL-3.0-or-later.

https://github.com/gnuradio/volk/commit/65b16cebefaf0b62df71cdb4942ea5c187247bbb#commitcomment-58909254

Signed-off-by: Johannes Demel <demel@uni-bremen.de>