lgpl: Add Mathieu to the list
We received the following statement

I, Mathieu Rene, hereby resubmit all my contributions to the VOLK project and repository under the terms of the LGPL-3.0-or-later. My GitHub handle mrene. I hereby agree that contributions made by me in the past, to previous versions of VOLK, may be re-used for inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPL-3.0-or-later.

https://github.com/gnuradio/volk/commit/1df5f16a10b854733a5a88a92c13a69c2d5b9bb3#commitcomment-58926589

Signed-off-by: Johannes Demel <demel@uni-bremen.de>