lgpl: Add Jam to the list
We received the following statement

I, Jam M. Hernandez Quiceno, hereby resubmit all my contributions to the VOLK project and repository under the terms of the LGPL-3.0-or-later. My GitHub handle JamMarHer. I hereby agree that contributions made by me in the past, to previous versions of VOLK, may be re-used for inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPL-3.0-or-later.

https://github.com/gnuradio/volk/commit/e655e2b8ea26f50020d2ef8112b9b74709d45ad0#commitcomment-59598504

Signed-off-by: Johannes Demel <demel@uni-bremen.de>