lgpl: Add Steven to the list
We received approval for the following statement

I, Steven Behnke, hereby resubmit all my contributions to the VOLK project and repository under the terms of the LGPL-3.0-or-later. My GitHub handle is `sbehnke`. I hereby agree that contributions made by me in the past, to previous versions of VOLK, may be re-used for inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPL-3.0-or-later.

https://github.com/gnuradio/volk/commit/79db96fa733a63d335549288b9b7dafe8b6fed3c#commitcomment-58894448

Signed-off-by: Johannes Demel <demel@uni-bremen.de>