lgpl: Add Nick to the list
We received approval for the following statement

I, Nick Foster, hereby resubmit all my contributions to the VOLK project and repository under the terms of the LGPL-3.0-or-later. My GitHub handle is `bistromath`. I hereby agree that contributions made by me in the past, to previous versions of VOLK, may be re-used for inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPL-3.0-or-later.

https://github.com/gnuradio/volk/commit/b0b9615e4e5d38c0d8d6bcc06ccefe08682ec352#commitcomment-58895312

Signed-off-by: Johannes Demel <demel@uni-bremen.de>