lgpl: Add Abhishek to the list
We received approval for the following statement

I, Abhishek Bhowmick, hereby resubmit all my contributions to the VOLK project and repository under the terms of the LGPL-3.0-or-later. My GitHub handle is `abhowmick22`. I hereby agree that contributions made by me in the past, to previous versions of VOLK, may be re-used for inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPL-3.0-or-later.

https://github.com/gnuradio/volk/commit/83cfe7b97e050af0f870a4a2e2059efe1efa22d1#commitcomment-60573221

Signed-off-by: Johannes Demel <demel@uni-bremen.de>