python: Switch to sysconfig module
The `distutils` module is deprecated and will be removed in Python 3.12.
We follow GNU Radio's example as well as the recommendations in PEP 632
and switch from `distutils.sysconfig` to `sysconfig`. Along with that
change comes one fewer package dependency on Debian. We no longer need
`python3-distutils` as well as an improvement to better distinguish
between `dist-packages` and `site-packages`.

Signed-off-by: Johannes Demel <demel@uni-bremen.de>