cpu_features: Make cpu_features submodule optional
Since cpu_features is available as a package in recent distributions,
we'd like to use the distro provided package.

We try to find the CMake CpuFeatures package and build against it, if we
can't find one, we try to use the git submodule. If both options fail,
we emit an error message.

Fix #556

Signed-off-by: Johannes Demel <demel@uni-bremen.de>