relicensing: Add Elliot to the list
We received the following statement:

I, Elliot Briggs, hereby resubmit all my contributions to the VOLK project and repository under the terms of the LGPL-3.0-or-later. My email addresses used for contributions are: elliot.briggs@gmail.com, ... .   I hereby agree that contributions made by me in the past, to previous versions of VOLK, may be re-used for inclusion in VOLK 3. I understand that VOLK 3 will be relicensed under LGPL-3.0-or-later.

Signed-off-by: Johannes Demel <demel@uni-bremen.de>