ci: Fix CI to use later distros
Some CI starts to fail depending on the underlying system. This commit
fixes this behavior. Also, later compilers are added. Though, for now
Clang14 on aarch64 is allowed to fail because of a newly discovered log2
accuracy issue. We might need to increase test tolerance here.

All test configurations should run faster now because we use `-j` to
speed up compilation. This defaults to 2 on Linux and Win. On MacOS we
use `-j3`. These are the default GHA runner sizes.
This option is not used to run tests to make sure it is as simple as
possible to read the test result output.

We should discuss when to remove CI that relies on GCC < 9.0 and Clang < 10.0
after next April. These are the default compilers in Ubuntu 20.04. Anything older
than this should be considered obsolete.

Signed-off-by: Johannes Demel <demel@uni-bremen.de>