cmake: Fix 64bit host CPU detection
In cases where we don't cross-compile, we might want to detect if a CPU
is 32bit or 64bit. CMake provides functionality for this case starting
in CMake 3.10. Let's use it.
Ubuntu 20.04 uses CMake 3.16. From the top of my head, this is the
oldest supported distribution. Debian buster ships with CMake 3.13.

Signed-off-by: Johannes Demel <jdemel@gnuradio.org>